refactor: remove common components
