+++ web/src/beta/features/Editor/Map/InspectorPanel/index.tsx
@@ -1,6 +1,5 @@
 import { FC, useMemo } from "react";
 
-import SceneSettings from "@reearth/beta/features/Editor/common/Settings";
 import { Panel, PanelProps } from "@reearth/beta/ui/layout";
 import { useT } from "@reearth/services/i18n";
 
@@ -8,6 +7,7 @@ import { useMapPage } from "../context";
 
 import LayerInspector from "./LayerInspector";
 import LayerStyleEditor from "./LayerStyleEditor";
+import SceneSettings from "./SceneSettings";
 
 type Props = Pick<PanelProps, "showCollapseArea" | "areaRef">;
